# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/overview": {
      "get": {
        "operationId": "OverviewController_findAll",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "idclient",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "situation",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OverviewResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar overview (acompanhamento geral)",
        "tags": [
          "Overview"
        ]
      }
    },
    "/overview/{id}": {
      "put": {
        "operationId": "OverviewController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOverviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Overview atualizado com sucesso"
          },
          "400": {
            "description": "Transição de status inválida ou dados inválidos"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar overview do cliente",
        "tags": [
          "Overview"
        ]
      }
    },
...
  },
  "info": {
    "title": "API - Controle de Vistorias e Entregas",
    "description": "Documentação da API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "InspectionRejectionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "prevision_date": {
            "type": "object",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "status",
          "prevision_date",
          "created_at",
          "updated_at"
        ]
      },
      "InspectionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "datetime": {
            "type": "string"
          },
          "inspector": {
            "type": "object",
            "nullable": true
          },
          "mobuss": {
            "type": "boolean"
          },
          "status": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "object",
            "nullable": true
          },
          "rejections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InspectionRejectionDto"
            }
          }
        },
        "required": [
          "id",
          "datetime",
          "inspector",
          "mobuss",
          "status",
          "created_at",
          "updated_at",
          "rejections"
        ]
      },
      "OverviewResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PENDENTE",
              "EM ANDAMENTO",
              "LIBERADA"
            ],
            "example": "PENDENTE"
          },
          "situation": {
            "type": "string",
            "enum": [
              "ATIVO",
              "INATIVO"
            ],
            "example": "ATIVO"
          },
          "status_recente": {
            "type": "string",
            "enum": [
              "ACEITE",
              "AGUARDANDO",
              "CANCELADA",
              "RECUSA"
            ],
            "example": "AGUARDANDO",
            "description": "Status recente calculado"
          },
          "status_quality": {
            "type": "object",
            "example": "PENDENTE",
            "nullable": true
          },
          "status_delivery": {
            "type": "object",
            "example": "PENDENTE",
            "nullable": true
          },
          "data_register": {
            "type": "object",
            "nullable": true
          },
          "data_contact": {
            "type": "object",
            "nullable": true
          },
          "obs": {
            "type": "object",
            "nullable": true
          },
          "inspections": {
            "example": "AGUARDANDO",
            "description": "Status recente calculado a partir das vistorias/recusas",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InspectionDto"
            }
          }
        },
        "required": [
          "status",
          "situation",
          "status_recente",
          "status_quality",
          "status_delivery",
          "data_register",
          "data_contact",
          "obs",
          "inspections"
        ]
      },
      "UpdateOverviewDto": {
        "type": "object",
        "properties": {
          "status_quality": {
            "type": "string",
            "example": "OK"
          },
          "status_construction": {
            "type": "string",
            "example": "CONCLUÍDA"
          },
          "status_delivery": {
            "type": "string",
            "example": "OK"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDENTE",
              "EM ANDAMENTO",
              "LIBERADA"
            ]
          },
          "obs": {
            "type": "string",
            "example": "Observações atualizadas"
          },
          "situation": {
            "type": "string",
            "enum": [
              "ATIVO",
              "INATIVO"
            ]
          }
        }
      },
...
    }
  }
}